COMPLIANCE-193

Approved-by: Kaspar Pedersen

+++ src/main/java/dk/digitalidentity/controller/rest/AssetsRestController.java
@@ -20,6 +20,9 @@ import dk.digitalidentity.model.entity.S3Document;
 import dk.digitalidentity.model.entity.enums.DPIAReportReportApprovalStatus;
 import dk.digitalidentity.model.entity.enums.EmailTemplatePlaceholder;
 import dk.digitalidentity.model.entity.enums.EmailTemplateType;
+import dk.digitalidentity.security.RequireSuperuser;
+import dk.digitalidentity.security.RequireUser;
+import dk.digitalidentity.security.Roles;
 import dk.digitalidentity.service.DPIAResponseSectionAnswerService;
 import dk.digitalidentity.service.DPIAResponseSectionService;
 import dk.digitalidentity.service.DPIATemplateQuestionService;
@@ -40,6 +43,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -58,7 +63,6 @@ import dk.digitalidentity.model.dto.PageDTO;
 import dk.digitalidentity.model.entity.Asset;
 import dk.digitalidentity.model.entity.User;
 import dk.digitalidentity.model.entity.grid.AssetGrid;
-import dk.digitalidentity.security.RequireUser;
 import dk.digitalidentity.service.AssetService;
 import dk.digitalidentity.service.UserService;
 import dk.digitalidentity.util.ReflectionHelper;
@@ -93,14 +97,13 @@ public class AssetsRestController {
     private final ApplicationEventPublisher eventPublisher;
     private final AssetOversightDao assetOversightDao;
 
-
     @PostMapping("list")
 	public PageDTO<AssetDTO> list(@RequestParam(name = "search", required = false) final String search,
                                   @RequestParam(name = "page", required = false, defaultValue = "0") final Integer page,
                                   @RequestParam(name = "size", required = false, defaultValue = "50") final Integer size,
                                   @RequestParam(name = "order", required = false) final String order,
                                   @RequestParam(name = "dir", required = false) final String dir) {
-		Sort sort = null;
+        Sort sort = null;
 		if (StringUtils.isNotEmpty(order) && containsField(order)) {
 			final Sort.Direction direction = Sort.Direction.fromOptionalString(dir).orElse(Sort.Direction.ASC);
 			sort = Sort.by(direction, order);
@@ -118,8 +121,10 @@ public class AssetsRestController {
 			assets = assetGridDao.findAll(sortAndPage);
 		}
 
-		return new PageDTO<>(assets.getTotalElements(), mapper.toDTO(assets.getContent()));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return new PageDTO<>(assets.getTotalElements(), mapper.toDTO(assets.getContent(), authentication.getAuthorities().stream().anyMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)), authentication.getPrincipal().toString()));
 	}
+
     @PostMapping("list/{id}")
     public PageDTO<AssetDTO> list(@PathVariable(name = "id", required = true) final String uuid,
                                 @RequestParam(name = "search", required = false) final String search,
@@ -129,6 +134,10 @@ public class AssetsRestController {
                                 @RequestParam(name = "dir", required = false) final String dir) {
         Sort sort = null;
         final User user = userService.findByUuid(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !authentication.getPrincipal().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (StringUtils.isNotEmpty(order) && containsField(order)) {
             final Sort.Direction direction = Sort.Direction.fromOptionalString(dir).orElse(Sort.Direction.ASC);
             sort = Sort.by(direction, order);
@@ -146,7 +155,7 @@ public class AssetsRestController {
             assets = assetGridDao.findAllByResponsibleUserUuidsContaining(user.getUuid(), sortAndPage);
         }
 
-        return new PageDTO<>(assets.getTotalElements(), mapper.toDTO(assets.getContent()));
+        return new PageDTO<>(assets.getTotalElements(), mapper.toDTO(assets.getContent(), authentication.getAuthorities().stream().anyMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)), authentication.getPrincipal().toString()));
     }
 
     @PutMapping("{id}/setfield")
@@ -154,6 +163,10 @@ public class AssetsRestController {
                                         @RequestParam(value = "value", required = false) final String value) {
         canSetFieldGuard(fieldName);
         final Asset asset = assetService.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !asset.getResponsibleUsers().stream().map(User::getUuid).toList().contains(authentication.getPrincipal().toString())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         ReflectionHelper.callSetterWithParam(Asset.class, asset, fieldName, value);
         assetService.save(asset);
     }
@@ -163,11 +176,16 @@ public class AssetsRestController {
         @RequestParam(value = "value", required = false) final String value) {
         canSetFieldDPIAScreeningGuard(fieldName);
         final Asset asset = assetService.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !asset.getResponsibleUsers().stream().map(User::getUuid).toList().contains(authentication.getPrincipal().toString())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         ReflectionHelper.callSetterWithParam(DataProtectionImpactAssessmentScreening.class, asset.getDpiaScreening(), fieldName, value);
         assetService.save(asset);
     }
 
     record DPIASetFieldDTO(long id, String fieldName, String value) {}
+    @RequireSuperuser
     @PutMapping("dpia/schema/section/setfield")
     public void setDPIASectionField( @RequestBody final DPIASetFieldDTO dto) {
         canSetDPIASectionFieldGuard(dto.fieldName);
@@ -176,30 +194,35 @@ public class AssetsRestController {
         dpiaTemplateSectionService.save(dpiaTemplateSection);
     }
 
+    @RequireSuperuser
     @PostMapping("dpia/schema/section/{id}/up")
     public ResponseEntity<?> reorderUp(@PathVariable("id") final long id) {
         reorderSections(id, false);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @RequireSuperuser
     @PostMapping("dpia/schema/section/{id}/down")
     public ResponseEntity<?> reorderDown(@PathVariable("id") final long id) {
         reorderSections(id, true);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @RequireSuperuser
     @PostMapping("dpia/schema/question/{id}/up")
     public ResponseEntity<?> reorderQuestionUp(@PathVariable("id") final long id) {
         reorderQuestions(id, false);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @RequireSuperuser
     @PostMapping("dpia/schema/question/{id}/down")
     public ResponseEntity<?> reorderQuestionDown(@PathVariable("id") final long id) {
         reorderQuestions(id, true);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @RequireSuperuser
     @DeleteMapping("dpia/schema/question/{id}/delete")
     public ResponseEntity<?> deleteQuestion(@PathVariable("id") final long id) {
         final DPIATemplateQuestion question = dpiaTemplateQuestionService.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
@@ -213,10 +236,15 @@ public class AssetsRestController {
     public ResponseEntity<?> deleteOversight(@PathVariable("oversightId") Long oversightId) {
         final AssetOversight assetOversight = assetService.getOversight(oversightId)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER) && assetOversight.getResponsibleUser().getUuid().equals(authentication.getPrincipal().toString()))) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         assetOversightDao.delete(assetOversight);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    @RequireSuperuser
     @Transactional
     @DeleteMapping("{assetId}/subsupplier/{subSupplierId}")
     public ResponseEntity<?> subSupplierDelete(@PathVariable("subSupplierId") final Long subSupplierId,
@@ -232,6 +260,10 @@ public class AssetsRestController {
     @PutMapping("{assetId}/dpia/response/setfield")
     public void setDPIAResponseField( @RequestBody final DPIASetFieldDTO dto, @PathVariable final long assetId) throws IOException {
         final Asset asset = assetService.findById(assetId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !asset.getResponsibleUsers().stream().map(User::getUuid).toList().contains(authentication.getPrincipal().toString())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (asset.getDpia() == null) {
             DPIA dpia = new DPIA();
             dpia.setAsset(asset);
@@ -289,6 +321,10 @@ public class AssetsRestController {
     @PutMapping("{assetId}/dpia/setfield")
     public void setDPIASectionField( @RequestBody final DPIASetFieldDTO dto, @PathVariable long assetId) {
         Asset asset = assetService.findById(assetId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !asset.getResponsibleUsers().stream().map(User::getUuid).toList().contains(authentication.getPrincipal().toString())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (dto.fieldName.equals("conclusion")) {
             asset.getDpia().setConclusion(dto.value);
         } else if (dto.fieldName.equals("checkedThreatAssessmentIds")) {
@@ -302,6 +338,10 @@ public class AssetsRestController {
     @PostMapping("{id}/mailReport")
     public ResponseEntity<?> mailReport(@PathVariable final long id, @RequestBody final MailReportDTO dto) throws IOException {
         Asset asset = assetService.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication.getAuthorities().stream().noneMatch(r -> r.getAuthority().equals(Roles.SUPERUSER)) && !asset.getResponsibleUsers().stream().map(User::getUuid).toList().contains(authentication.getPrincipal().toString())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         final User responsibleUser = userService.findByUuid(dto.sendTo).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Den valgte bruger kunne ikke findes, og rapporten kan derfor ikke sendes."));
         if (responsibleUser.getEmail() == null || responsibleUser.getEmail().isBlank()) {
             return new ResponseEntity<>("Den valgte bruger har ikke nogen email tilknyttet, og rapporten kan derfor ikke sendes.", HttpStatus.BAD_REQUEST);
